Harvest WoS publications en masse
Chunk it up by date range.
