+++ lib/rialto/etl/service_client/retriable_connection_factory.rb
@@ -25,7 +25,7 @@ module Rialto
                                        retry_statuses: retry_statuses
 
             connection.ssl.update(verify: true, verify_mode: OpenSSL::SSL::VERIFY_PEER)
-            connection.adapter :net_http_persistent
+            connection.adapter :net_http
             connection.options.timeout = 500
             connection.options.open_timeout = 10
           end
